requirements.txt
# Requirements for the iarbackend itself
django>=2.0
psycopg2
# explicitly specify django-automationcommon's git repo since changes in
# automationcommon tend to be "ad hoc" and may need testing here without a
# corresponding pypi release. Recall that git branched may be explicitly given
# in the VCS URL.
git+https://github.com/uisautomation/django-automationcommon.git@master#egg=django-automationcommon
django-ucamwebauth>=1.4.7
django-multiselectfield
djangorestframework
#django rest framework documentation dependencies
coreapi
pygments
markdown
django-cors-headers
drf-yasg[validation]

# For managing OAuth2 tokens
oauthlib
requests-oauthlib

# For an improved python manage.py shell experience
ipython

# So that tests may be run within the container
tox

# Serving
gunicorn
